tests fix

+++ substrate-sdk-android/src/test/java/io/novasama/substrate_sdk_android/common/Math.kt
@@ -3,7 +3,7 @@ package io.novasama.substrate_sdk_android.common
 /**
  * Complexity: O(n * log(n))
  */
-fun List<Double>.median(): Double = sorted().let {
+fun List<Long>.median(): Long = sorted().let {
     val middleRight = it[it.size / 2]
     val middleLeft = it[(it.size - 1) / 2] // will be same as middleRight if list size is odd
